+++ editor/vite.config.js
@@ -12,6 +12,7 @@ export default defineConfig({
       entry: 'src/main.js',
       formats: ['es'],
       fileName: () => 'editor.js',
+      cssFileName: 'editor',
     },
   },
 });